+++ NEWS.md
@@ -1,4 +1,4 @@
-# bench (development version)
+# bench 1.1.0
 
 ## New features